Updated database access to support other methods of connecting

+++ mtdrworkshop/backend/src/main/java/com/oracle/todoapp/TodoItemStorage.java
@@ -1,5 +1,5 @@
 /*
-## MyToDoReact version 2.0.0
+## MyToDoReact version 2.1.0
 ##
 ## Copyright (c) 2021 Oracle, Inc.
 ## Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
@@ -18,10 +18,9 @@ import java.util.concurrent.CompletionStage;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import com.oracle.database.AppDatasource;
 import io.helidon.config.Config;
-import oracle.ucp.jdbc.PoolDataSource;
-import oracle.ucp.jdbc.PoolDataSourceFactory;
-//import io.helidon.dbclient.jdbc;
+
 /*
  * This class takes care of the storage of the todo items. It uses an Autonomous Database
  * from the Oracle Cloud (ATP). The following table is used to store the todo items:
@@ -38,47 +37,14 @@ import oracle.ucp.jdbc.PoolDataSourceFactory;
  */
 
 class TodoItemStorage {
-  // grabbing the dbpassword from the kubernetes secret, added by peter song
-  static String pwSecretFromK8s = System.getenv("dbpassword").trim();
+
   private final static Logger LOGGER = Logger.getLogger(TodoItemStorage.class.getName());
 
-  private final PoolDataSource pool;
+  private final AppDatasource ads;
 
-  // singleton repository:
-  private static TodoItemStorage storage = null;
-
-  synchronized static TodoItemStorage get(Config config) {
-    if(storage == null) {
-      try {
-        storage = new TodoItemStorage(config);
-      } catch (SQLException ex) {
-        LOGGER.log(Level.SEVERE, ()->config.toString());
-        LOGGER.log(Level.SEVERE, ex, ()->"in TodoItemRepository.get(...)");
-      }
-    }
-    return storage;
-  }
-
-  private TodoItemStorage(Config config)  throws SQLException {
-    LOGGER.log(Level.CONFIG, ()->config.toString());
-    // trying this in place of "url and user and dbpasswor"
-    //String password = config.get("password").asString().get();
-    String url = config.get("url").asString().get();
-    String user = config.get("user").asString().get();
-    System.out.printf("Using url: %s%n", url);
-    pool = PoolDataSourceFactory.getPoolDataSource();
-    pool.setURL(url);
-    //pool.setUser(user);
-    pool.setUser("TODOUSER");
-    pool.setPassword(pwSecretFromK8s);
-    pool.setInactiveConnectionTimeout(60);
-    pool.setConnectionFactoryClassName("oracle.jdbc.pool.OracleDataSource");
-    pool.setMaxStatements(10);
-    // In this application, we don't set any init, min or max size in UCP. We
-    // also don't start the pool explicitly. This means that the very first 
-    // connection request will start the pool. The default maximum pool size
-    // is MAX_INT which isn't appropriate and should be configured properly in
-    // production.
+  TodoItemStorage(Config config) {
+    LOGGER.log(Level.CONFIG, config::toString);
+    ads = AppDatasource.get(config);
   }
 
   /**
@@ -100,8 +66,8 @@ class TodoItemStorage {
     LOGGER.fine("all");
     ArrayList<TodoItem> list = new ArrayList<TodoItem>();
     try (
-      Connection conn = pool.getConnection();
-      PreparedStatement pstmt = conn.prepareStatement("SELECT id, description, creation_ts, done FROM todoitem ORDER BY creation_ts DESC");
+      Connection conn = ads.getConnection();
+      PreparedStatement pstmt = conn.prepareStatement("SELECT id, description, creation_ts, done FROM TODOOWNER.todoitem ORDER BY creation_ts DESC");
       ResultSet rs = pstmt.executeQuery();
     ) {
       while(rs.next()) {
@@ -124,8 +90,8 @@ class TodoItemStorage {
     LOGGER.fine(()->"getById("+id+")");
     TodoItem ret = null;
     try (
-        Connection conn = pool.getConnection();
-        PreparedStatement pstmt = conn.prepareStatement("SELECT id, description, creation_ts, done FROM todoitem WHERE id=?");
+        Connection conn = ads.getConnection();
+        PreparedStatement pstmt = conn.prepareStatement("SELECT id, description, creation_ts, done FROM todoowner.todoitem WHERE id=?");
         ){
       pstmt.setInt(1, id);
       try (ResultSet rs = pstmt.executeQuery();) {
@@ -163,11 +129,11 @@ class TodoItemStorage {
     LOGGER.fine("save("+item.toString()+")");
     int idInt = item.getId();
     try (
-      Connection conn = pool.getConnection();
+      Connection conn = ads.getConnection();
     ){
       if(idInt >=0 ) {
         try(
-          PreparedStatement pstmt = conn.prepareStatement("UPDATE todoitem SET description=?, done=? WHERE id=?");
+          PreparedStatement pstmt = conn.prepareStatement("UPDATE todoowner.todoitem SET description=?, done=? WHERE id=?");
         ){
           pstmt.setString(1, item.getDescription());
           pstmt.setInt(2, item.isDone()?1:0);
@@ -178,7 +144,7 @@ class TodoItemStorage {
         return this.getById(idInt);
       } else {
         try (
-          PreparedStatement pstmt = conn.prepareStatement("INSERT INTO todoitem (description) VALUES (?)", new String[]{"id"});
+          PreparedStatement pstmt = conn.prepareStatement("INSERT INTO todoowner.todoitem (description) VALUES (?)", new String[]{"id"});
         ){
           pstmt.setString(1, item.getDescription());
 
@@ -211,8 +177,8 @@ class TodoItemStorage {
     LOGGER.fine(()->"deleteById(" + id + ")");
     int ret = 0;
     try (
-        Connection conn = pool.getConnection();
-        PreparedStatement pstmt = conn.prepareStatement("DELETE FROM todoitem WHERE id=?");
+        Connection conn = ads.getConnection();
+        PreparedStatement pstmt = conn.prepareStatement("DELETE FROM todoowner.todoitem WHERE id=?");
       ){
       pstmt.setInt(1, id);
       ret = pstmt.executeUpdate();
